refactor(cli): extract Petdex mascot cluster into CLIPetMixin (cli.py god-file slice R2)
Move the 13-member Petdex mascot cluster (banner + _PET_FRAME_INTERVAL,
_PET_CFG_INTERVAL + 11 _pet*/_on_reaction methods, cli.py window 5613-5829
at pin aec331899e) byte-verbatim into hermes_cli/cli_pet_mixin.py as
CLIPetMixin, appended last to HermesCLI's base tuple. Identity-preserving
seam: all members resolve via MRO; no __init__ change; zero caller edits
(all 7 out-of-cluster call sites are same-class). Part of #55136
